reordering removed and Clip issues are prevented
