+++ models/experimental/stable_diffusion_xl_base/demo/demo.py
@@ -52,7 +52,7 @@ def run_demo_inference(
     needed_padding = (batch_size - len(prompts) % batch_size) % batch_size
     prompts = prompts + [""] * needed_padding
 
-    guidance_scale = 5.0
+    guidance_scale = 8.0
 
     # 0. Set up default height and width for unet
     height = 1024
@@ -201,16 +201,6 @@ def run_demo_inference(
     profiler.end("encode_prompts")
 
     profiler.start("prepare_latents")
-    # Reorder all_embeds to prepare for splitting across devices
-    items_per_core = len(all_embeds) // batch_size  # this will always be a multiple of batch_size because of padding
-
-    if batch_size > 1:  # If batch_size is 1, no need to reorder
-        reordered = []
-        for i in range(batch_size):
-            for j in range(items_per_core):
-                index = i + j * batch_size
-                reordered.append(all_embeds[index])
-        all_embeds = reordered
 
     prompt_embeds, negative_prompt_embeds, pooled_prompt_embeds, negative_pooled_prompt_embeds = zip(*all_embeds)
